Rename from Song Refresh Button to Mediocre Mapper Pull
Though I used Song Refresh Button as a starting point because it
triggered song refresh, I didn't realize when I started that it called
out to SongCore to refresh songs. Alppuccino and I discussed it and
decided that pulling from Mediocre Mapper is a niche use case and it
would make more sense to have separate plugins.